Tourism Website Based on Laravel, PHP.
RTG (Ready to Go) Travels is a travel agency devoted to providing the best possible experience to its users. It is an online travel booking website that provides comprehensive information about the   offered destinations, the best itineraries, top-notch amenities, and an unforgettable experience all while giving the users the ability to book tours online without any hassle. For the cs350 Web Engineering project, the aim of the group was to design and build a cutting-edge, professional website for this travel agency. Using this website, the users will be able to sign up, book tours from a list of pre-prepared travel packages or customize a travel package that is tailored to their needs. Moreover, the travelers can explore the website to gain assistance for planning their next trip. The responsive script and layout have been designed through HTML, CSS and Bootstrap v.5.2.3. The server-side programming has been achieved through php and Laravel with the database records stored in a mysql database.